src/graph_bridges/models/ising/boltzman_machines.py
import torch


# Define the fully visible Boltzmann machine model
class FullyVisibleBoltzmannMachine(torch.nn.Module):
    def __init__(self, num_visible, num_hidden):
        super(FullyVisibleBoltzmannMachine, self).__init__()
        self.num_visible = num_visible
        self.num_hidden = num_hidden
        self.W = torch.randn(num_visible, num_hidden)
        self.bv = torch.randn(num_visible)
        self.bh = torch.randn(num_hidden)

    def energy(self, v):
        """Calculate the energy of the Boltzmann machine for a given visible state"""
        hidden_term = torch.sum(torch.log(1 + torch.exp(self.bh + torch.matmul(v, self.W))))
        visible_term = torch.sum(torch.mul(v, self.bv))
        return -hidden_term - visible_term

    def forward(self, v):
        """Calculate the probability of a hidden state given a visible state"""
        return torch.sigmoid(self.bh + torch.matmul(v, self.W))

    def sample(self, num_samples=1, init_state=None):
        """Generate samples from the Boltzmann machine using Gibbs sampling"""
        if init_state is None:
            init_state = torch.zeros(num_samples, self.num_visible)
        curr_state = init_state
        for _ in range(num_samples):
            # Sample a new hidden state from the current visible state
            hidden_probs = self.forward(curr_state)
            hidden_states = (torch.rand(num_samples, self.num_hidden) < hidden_probs).float()
            # Sample a new visible state from the new hidden state
            visible_probs = torch.sigmoid(self.bv + torch.matmul(hidden_states, self.W.t()))
            visible_states = (torch.rand(num_samples, self.num_visible) < visible_probs).float()
            curr_state = visible_states
        return curr_state


if __name__=="__main__":
    number_of_paths = 100
    number_of_spins = 10
    # Create a fully visible Boltzmann machine with 4 visible units and 2 hidden units
    boltzman_machine = FullyVisibleBoltzmannMachine(num_visible=number_of_spins, num_hidden=2)
    # Generate 5 samples from the Boltzmann machine
    boltzman_machine_sample = boltzman_machine.sample(num_samples=number_of_paths)